FDMS/requirements.txt
pandas==2.0.0
sqlalchemy==1.4.39
psycopg2-binary==2.9.5
boto3==1.26.137
python-dotenv==1.0.0 